// Kudzu/gis/Datum.d.ts
export declare const DatumWGS_84: {
    FalseNorthing: number;
    equatorialRadius: number;
    pointScaleFactor: number;
    E0: number;
    A: number;
    flattening: number;
    e: number;
    esq: number;
    e0sq: number;
    alpha1: number;
    alpha2: number;
    alpha3: number;
    alpha4: number;
    beta: number[];
    delta: number[];
};